Almost done moving lobby manager into partials
