+++ Multiplayer/Services/Steam/Lobby/LobbyService.Guest.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Threading.Tasks;
+using GodotPeer2PeerSteamCSharp.Core.Lobby;
 using Steamworks;
 
 namespace GodotPeer2PeerSteamCSharp.Services.Steam.Lobby;
