+++ Scripts/UI/Lobby/LobbyChat.cs
@@ -1,4 +1,5 @@
 using Godot;
+using GodotPeer2PeerSteamCSharp.Core.Lobby;
 
 public partial class LobbyChat : Node
 {
