+++ Types/Scene/SceneRegistry.cs
@@ -1,4 +1,4 @@
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
 namespace GodotPeer2PeerSteamCSharp.Types.Scene;
