+++ project.godot
@@ -19,7 +19,7 @@ config/icon="res://icon.svg"
 
 Logger="*res://Core/Logger.cs"
 EventBus="*res://Core/EventBus.cs"
-LobbyManager="*res://Core/LobbyManager.cs"
+LobbyManager="*res://Core/Lobby/LobbyManager.cs"
 TransportManager="*res://Core/TransportManager.cs"
 InputManager="*res://Core/InputManager.cs"
 SceneManager="*res://Core/SceneManager/SceneManager.cs"
